backend/requirements.txt
alembic==1.16.4
annotated-types==0.7.0
anyio==4.10.0
blinker==1.9.0
Brotli==1.1.0
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
cors==1.0.1
cryptography==45.0.6
cssselect2==0.8.0
deprecation==2.1.0
filelock==3.18.0
Flask==3.1.1
flask-cors==6.0.1
Flask-Dance==7.1.0
Flask-Mail==0.10.0
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
fonttools==4.58.4
future==1.0.0
gevent==25.5.1
google-auth==2.40.3
greenlet==3.2.3
gunicorn==23.0.0
h11==0.16.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.2
oauthlib==3.3.1
packaging==25.0
pdf2image==1.17.0
pillow==11.2.1
postgrest==1.1.1
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pydyf==0.11.0
PyJWT==2.10.1
pyphen==0.17.2
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
realtime==2.7.0
requests==2.32.4
requests-file==2.1.0
requests-oauthlib==2.0.0
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.41
storage3==0.12.1
StrEnum==0.4.15
supabase==2.18.1
supabase_auth==2.12.3
supabase_functions==0.10.1
tinycss2==1.4.0
tinyhtml5==2.0.0
tldextract==5.3.0
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.4.0
URLObject==3.0.0
weasyprint==65.1
webencodings==0.5.1
websockets==15.0.1
Werkzeug==3.1.3
zope.event==5.0
zope.interface==7.2
zopfli==0.2.3.post1
